// src/types/prismic-types.ts
type Simplify<T> = {
    [KeyType in keyof T]: T[KeyType];
};
/** Content for catalogue documents */
interface CatalogueDocumentData {
    /**
     * title field in *catalogue*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: catalogue.title
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    title: prismicT.KeyTextField;
    /**
     * year field in *catalogue*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: catalogue.year
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    year: prismicT.KeyTextField;
    /**
     * price field in *catalogue*
     *
     * - **Field Type**: Text
     * - **Placeholder**: $999
     * - **API ID Path**: catalogue.price
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    price: prismicT.KeyTextField;
    /**
     * images field in *catalogue*
     *
     * - **Field Type**: Group
     * - **Placeholder**: *None*
     * - **API ID Path**: catalogue.images[]
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/group
     *
     */
    images: prismicT.GroupField<Simplify<CatalogueDocumentDataImagesItem>>;
    /**
     * specs field in *catalogue*
     *
     * - **Field Type**: Rich Text
     * - **Placeholder**: *None*
     * - **API ID Path**: catalogue.specs
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
     *
     */
    specs: prismicT.RichTextField;
    /**
     * description field in *catalogue*
     *
     * - **Field Type**: Rich Text
     * - **Placeholder**: *None*
     * - **API ID Path**: catalogue.description
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
     *
     */
    description: prismicT.RichTextField;
    /**
     * more details field in *catalogue*
     *
     * - **Field Type**: Rich Text
     * - **Placeholder**: *None*
     * - **API ID Path**: catalogue.more_details
     * - **Tab**: Main
     * - **Documentation**: https://prismic.io/docs/core-concepts/rich-text-title
     *
     */
    more_details: prismicT.RichTextField;
}
/**
 * Item in catalogue → images
 *
 */
export interface CatalogueDocumentDataImagesItem {
    /**
     * image field in *catalogue → images*
     *
     * - **Field Type**: Image
     * - **Placeholder**: *None*
     * - **API ID Path**: catalogue.images[].image
     * - **Documentation**: https://prismic.io/docs/core-concepts/image
     *
     */
    image: prismicT.ImageField<never>;
    /**
     * caption field in *catalogue → images*
     *
     * - **Field Type**: Text
     * - **Placeholder**: *None*
     * - **API ID Path**: catalogue.images[].caption
     * - **Documentation**: https://prismic.io/docs/core-concepts/key-text
     *
     */
    caption: prismicT.KeyTextField;
}
/**
 * catalogue document from Prismic
 *
 * - **API ID**: `catalogue`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/core-concepts/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type CatalogueDocument<Lang extends string = string> = prismicT.PrismicDocumentWithUID<Simplify<CatalogueDocumentData>, "catalogue", Lang>;